+++ src/dataprotection/azext_dataprotection/manual/helpers.py
@@ -1005,7 +1005,7 @@ def convert_backup_instance_show_to_input(backup_instance):
                     'resourceProperties' in datasource_info and
                     isinstance(datasource_info['resourceProperties'], dict)):
                     if datasource_info['resourceProperties'].get('objectType') is None:
-                        # Set resourceProperties to None when objectType is null to avoid schema validation error
+                        # Cleaning up resourceProperties when objectType is null to avoid schema validation error
                         del backup_instance['properties'][datasource_property]['resourceProperties']
     return backup_instance