+++ src/dataprotection/HISTORY.rst
@@ -2,9 +2,9 @@
 
 Release History
 ===============
-1.7.2
+1.8.0
 +++++
-* `az dataprotection backup-instance update`: New parameters: `--backup-configuration` to update AKS datasource parameters.
+* `az dataprotection backup-instance update`: New parameter: `--backup-configuration` to update AKS datasource parameters.
 * Fix in `helpers.py` to correctly prepare/normalize AKS backup-configuration payloads passed via the CLI.
 
 1.7.1